Migrate nearest neighborhood for rectangular grids
This requires two helper functions which could be moved to the grid ...

+++ dune/utopia/core/types.hh
@@ -1,6 +1,8 @@
 #ifndef UTOPIA_CORE_TYPES_HH
 #define UTOPIA_CORE_TYPES_HH
 
+#include <cstdint>
+
 namespace Utopia
 {
     
@@ -61,13 +63,13 @@ using entity_t = typename Container::value_type::element_type;
 // -- Types introduces with the new CellManager ------------------------------
 
 /// Type for indices
-using IndexType = std::size_t;
+using IndexType = std::uint_fast32_t;
 
 /// Type for container of indices
 using IndexContainer = std::vector<IndexType>;
 
-/// Type for dimensions
-using DimType = unsigned short;
+/// Type for dimension numbers, i.e.: small unsigned integers
+using DimType = uint_fast8_t;
 
 /// Type for grid shape
 template<DimType dim>
